그룹 내에서 연락처 정보 업데이트 분리
